Add a simple thread executor
