+++ judge/core/test_something.py
@@ -8,4 +8,4 @@ from random import randint
 def test_something():
     for i in range(100):
         assert isinstance(i, int), 'it should be a string'
-        sleep(randint(1, 3))
+        #sleep(randint(1, 3))
